requirements.txt
aiohttp==3.7.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
Django==2.2
django-filter==2.3.0
django-mathfilters==1.0.0
djangorestframework==3.11.1
Fabric3==1.14.post1
idna==2.10
loguru==0.5.3
lxml==4.5.2
multidict==5.0.0
paramiko==2.7.2
Pillow==7.2.0
pycparser==2.20
PyNaCl==1.4.0
pyTelegramBotAPI==3.7.3
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.25.9
yarl==1.6.2
